SimpleChains

Documentation for SimpleChains.

SimpleChains.AbstractPenalty

SimpleChains.Dropout

SimpleChains.FrontLastPenalty

Base.front

SimpleChains.valgrad!

SimpleChains.AbstractPenalty — Type

AbstractPenalty

The AbstractPenalty interface requires supporting the following methods:

getchain(::AbstractPenalty)::SimpleChain returns a SimpleChain if it is carrying one.

apply_penalty(::AbstractPenalty, params)::Number returns the penalty

apply_penalty!(grad, ::AbstractPenalty, params)::Number returns the penalty and updates grad to add the gradient.

SimpleChains.Dropout — Type

Dropout(p) # 0 < p < 1

Dropout layer.

When evaluated without gradients, it multiplies inputs by (1 - p). When evaluated with gradients, it randomly zeros p proportion of inputs.

SimpleChains.FrontLastPenalty — Type

FrontLastPenalty(SimpleChain, frontpen(λ₁...), lastpen(λ₂...))

Applies frontpen to all but the last layer, applying lastpen to the last layer instead. "Last layer" here ignores the loss function, i.e. if the last element of the chain is a loss layer, the then lastpen applies to the layer preceding this.

Base.front — Method

Base.front(c::SimpleChain)

Useful for popping off a loss layer.

SimpleChains.valgrad! — Method

Allowed destruction:

valgrad_layer! Accepts return of previous layer (B) and returns an ouput C. If an internal layer, allowed to destroy B (e.g. dropout layer).

pullback! Accepts adjoint of its return (C̄). It is allowed to destroy this. It is also allowed to destroy the previous layer's return B to produce B̄ (the C̄ it receives). Thus, the pullback is not allowed to depend on C, as it may have been destroyed in producing C̄.